implemented thread pool by following the official rust book's tutorial

+++ src/common/properties.rs
@@ -9,8 +9,8 @@ use std::error::Error;
 pub struct Properties {
     pub node_ips: Vec<Ipv4Addr>,
     pub server_port: u64,
-    pub dht_num_threads: u64,
-    pub dht_num_buckets: u64,
+    pub dht_num_threads: usize,
+    pub dht_num_buckets: usize,
     pub key_range: Vec<u64>,
     pub num_requests: u64
 }
